ai-agent/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
beautifulsoup4>=4.12.0
pgvector>=0.2.3
psycopg2-binary>=2.9.9
tiktoken>=0.5.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.3.2
requests>=2.31.0
aiohttp>=3.9.1
playwright>=1.41.0
selenium>=4.16.0
pydantic>=2.5.0